add graphviz output and random graph generation
Add the `src/graphviz.hpp` header which implements some helper functions
to output a graph in Graphviz format. This then allows the graph to be
rendered using dot, neato, and the like. In case an MST has been
determined, the edges belonging to the MST may also be highlighted.

Also add a helper function to generate an Erdös-Renyi graph based on a
desired number of edges and vertices. The graph may be disconnected or
contain parallel edges. Boost's implementation of Prim's algorithm seems
not to like that.

Add some scripts that generate rendered PDFs for convenience.