impl custom Prim and benchmark
Add custom implementation of Prim's algorithm which operates on the
vector-based graph representation. Add a performance timer that allows
for sections of the program to be timed. Benchmark the boost and custom
sequential implementation against each other. Current bottleneck lies in
generating the input graph from scratch. Generating the graph once on
disk should alleviate this issue.