impl graph loading from disk
Implement writing and reading graphs from disk. This allows graphs to be
generated offline and then loaded for the processing. The implementation
uses mmap to speed up the graph loading. Note that this pushes some IO
time into the MST execution, since the operating system will page in
blocks of the file on-demand, rather than upfront.