add second uniform pairs implementation
Add a second implementation of uniform pairs based on re-drawing of
random numbers. Instead of maintaining a precise list of gaps and using
that as a distribution, simply draw indices in the overall range of
potential target vertices and in case the connection has already been
made, draw another number. This implementation is much faster in the
case of a loosely connected graph, but will take very long to converge
in case of a highly connected graph due to the undirected random number
drawing.