add LFSR-based random sequence generator
Add an implementation of a pseudo-random maximum-length sequence
generator based on LFSRs. This is helpful for the creation of
Erdös-Renyi graphs as it allows long sequences of non-repeating pairs to
be generated, thus applying itself well to the G(n,p) model.

The implementation accepts an arbitrary range within which the generated
numbers will be. It rounds up to the next power of two and creates an
LFSR with the corresponding number of bits. Samples are then drawn from
that LFSR, discarding any that are above the range. In the worst case
half of the samples are discarded, such that two random numbers must be
drawn to generate one sample.

Add a test for the generator.