update draw-boost script
Turns out that the boost implementation of Prim's algorithm does not
find forests of MSTs, but simply the MST that goes through the very
first vertex in the set.

Adjust the `draw-boost.sh` script to produce a sample graph with three
separate subgraphs, only the first of which actually has its MST
created.